requirements.txt
blinker==1.3
Flask==0.10.1
Flask-Bootstrap==3.3.0.1
Flask-HTTPAuth==2.3.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
python-ldap==2.4.19
raven==5.2.0
Werkzeug==0.10.1
